Add Doxygen configuration and tie it into CMake.
By default, a "documentation" target (i.e. Visual Studio project under
VS) will be created if Doxygen is found, but will not be built
automatically.  To build the documentation, right-click it and select
Build in VS, run "make documentation" on the command line, etc.

To disable generating documentation altogether, set
SURGSIM_DOCUMENTATION to false.

To enable generating documentation during the build, set
SURGSIM_DOCUMENTATION_AUTOBUILD to true.  Note that this will cause it
to be (re)generated during each and every build.