Add licensing information for the project.
The LICENSE file is a direct copy of the Apache License, version 2.0:
    http://apache.org/licenses/LICENSE-2.0.txt

Source and other files should include a short reference to the
license; by the time you read this, there should be other source files
you can look at and copy from.

The license also makes provision for a NOTICE file, which exists to
collect required third-party attribution notices.  However, we
currently don't include any third-party code that requires such
notices, and the NOTICE file is not required.