Include GTest (v1.6.0) in the source tree.
We'll be using GTest for testing, but that requires either asking each
user to manually download and build it, or including it into our tree.
Note that you can't just install a pre-built library, even on Linux,
because distributing pre-built libs is discouraged by GTest's own FAQ:

    http://code.google.com/p/googletest/wiki/FAQ#Why_is_it_not_recommended_to_install_a_pre-compiled_copy_of_Goog

As a result, distributions like Ubuntu aren't providing it in library
form any more.  (Ubuntu does provide it in source form, but you can't
build or install it without superuser privileges, so it's not useful.)