Implement of groups of named values (poses, scalars...) for data exchange.
The NamedData template class can be used to provide a collection of
named values of some type, as specified by the template argument.  The
values can be accessed by name or by the corresponding index.  (Access
by index currently does not check whether the index really came from
the same data layout; we could add a safe way to do this in the
future.)

The DataGroup class provides a grouping of named values for some
commonly used types.

The data layout (i.e. name <-> index mapping(s)) for a NamedData or
DataGroup object is intentionally immutable; once the object is
initialized, the layout cannot be changed.  Moreover, objects can be
assigned only if the left side of the assignment is either
uninitialized (has no layout), or identical layout as the
right side.

Each of the two classes also has a corresponding ...Builder class,
which is used to initialize the data layout.