Fix Linux errors.
Note that the code no longer compiles with Boost 1.48, which is the
latest prepackaged Boost version in the (non-PPA) repositories for
Ubuntu 12.04 LTS (Precise).  In order to provide better diagnostics, I
added a requirement for Boost >= 1.51 to CMakeLists.  (I don't know
whether 1.49 or 1.50 would actually work, but we know 1.51 does.)