Fix the build for Linux.
As of C++11, you should be able to call std::vector's erase() with a
*const* iterator.  But apparently libstdc++ doesn't support that yet
as of gcc 4.7.2.

The other problem was a "feature" of CMake's FindGLUT code.

Note that I haven't checked if the example apps run on Linux, since I
don't have a Linux version of OpenHaptics available.  But compilation
against the OpenHaptics headers works, at least.