Simple Glut rendering for the MMVR2013_BiggerInputDemo using a new listener: SimpleSquareGlutWindow.
The square is drawn in white.
XYZ axes are drawn for the device pose in RGB.
A point is drawn at the tip point. If no button pressed, it is white. If button 0 is pressed, the R becomes 0.0; button 1 makes G 0.0; button 2 makes B 0.0; and button 3 decreases the brightness of the color by half.