Split the generic listener into an input consumer and an output producer.
This makes the flow of information more clear, and exposes the fact
that a device can only have one producer and an arbitrary number of
consumers.  Split and beefed up the tests, too.

Also, consistently use app-centric input/output language in the docs
(i.e. get rid of "app input which is the device output" nonsense).