Allow CMake to find "dot" for Doxygen via environment.
You can now set DOXYGEN_DOT_PATH environment variable to the Graphviz
binary directory path (if installed) to allow our CMake setup to find
dot.exe on Windows.  Out of the box, Graphviz versions newer than 2.21
are not detected.

Also include Doxyfile.in in the documentation project, making it easy
to open it in Visual Studio.