Allow Doxygen to use private members in collaboration diagrams.
With the default settings, collaboration diagrams for classes where
all data members are private (as most should be) look exactly
identical to inheritance diagrams.  That's pretty useless.

This changes a few settings in order to try to get more sensible
collaboration diagrams.  The alternative would be to turn off
collaboration diagrams altogether.