Use unique_ptr to opaque internal state, instead of a native pointer.
CommonDevice, PhantomDevice and PhantomManager use a struct to store
internal state without exposing the internals of what's stored.  We
used to hold a pointer to the struct, but as @hscheirich pointed out,
using a unique_ptr is better for both code robustness (the change
fixed at least one memory leak bug!) and exception safety.

Note that when a class C has a unique_ptr<T> member, the type T must
not be incomplete *when the destructor is being emitted by the
compiler*.  That means that if T is not defined in the header, you
need to define a (possibly empty) destructor for C *in its .cpp file*,
so I added one for CommonDevice.