Start with Identity in makeRigidTransform
Don't know if this effects anything, but makeRigidTransform returns a
transform, that internally has unitialized values on the bottom row. For
a rigid transform, these values might not come into play. But, it is
diconcerting to see them when debugging. This change initializes the
transform to identity and then sets the rotation and translation.
The bottom row is now correctly [0.0, 0.0, 0.0, 1.0].