Refactor CMake code to allow tests that are built but not run.
In some cases (notably for device code that requires specific
hardware), we want to be able to build tests by default, but not
actually run them unless the user says so.  This wasn't possible
before; if you built a unit test, it would get run too.

You can specify extra options to specify if a test will be run, e.g.:
    surgsim_add_unit_tests(myTestName RUN OFF)
Note that even if you say "RUN TRUE" etc, the test will not be run if
the global SURGSIM_TESTS_RUN flag is off, or if the test isn't built.

For symmetry, also allow turning off builds via the same mechanism
(BUILD OFF/ON/TRUE/FALSE).