Add unit tests for PhantomDevice and PhantomManager.
Note that the unit tests will NOT be run by default as a part of the
unit tests that run during the build, since the tests require hardware
to be connected in order to pass.  If you want to enable running them
during the build, turn on the SURGSIM_TESTS_RUN_DEVICE_PHANTOM option
in CMake.