Check for a rare problem with CMake's definition of MSVC_VERSION.
If building with MSVC, we now make sure MSVC_VERSION is defined.  It
almost always is.  But if you have cmake-gui open, and you delete
CMake's cache file in the build directory (e.g. because you're
deleting the entire build directory to ensure a clean build), and you
just click "Generate", the build will forget the definition of
MSVC_VERSION!

The problem is easy to fix-- just select "File" > "Delete Cache" in
cmake-gui, or restart cmake-gui.  But if you forget, you'll get weird
fallout from version checks being triggered incorrectly.

As far as I know, it's not possible to see this problem without a
persistent tool like cmake-gui or ccmake; the command line cmake
should be fine.