Move Physics Actors into Actors sub-directory.
Add initial Constraint, ConstraintData, and ConstraintImplementation classes.
A Constraint has some ConstraintData which is specific to the type of Constraint and 2 ConstraintImplementations, one for each side of the constraint.
Constraint::build() passes the ConstraintData to the 2 ConstraintImplementations' build() methods to build the constraint in the matrix.

Add initial Localization class.
A Localization stores a position on an Actor. The position can be calculated with calculatePosition().

Add license headers and run astyle on Physics code.