Add support for alternate handling of assertion failures.
The default is still to throw an AssertionFailure exception, but now
you can call AssertMessage::setFailureBehaviorToDebugger() and
assertion failures will to go straight into the debugger.

A convenient side effect is to mask all of the messages about throwing
from the AssertionFailure destructor.  (The reason for the warning is
still there-- using the exception-based assertion handling, we might
leak resources after an assertion failure.  But if an assertion has
failed, leaking a small amount of resources is probably the least of
your problems.)