Initial graphics framework.
The Graphics Manager derives from the Framework Manager and is added to the main Runtime to provide a thread for graphics update and rendering.
The Manager also manages the Graphics Scene and views of the scene with ViewComponents, each of which can be a separate window.
Actors make up the objects in the Scene.
For extra flexibility and to avoid coupling with specific graphics libraries and toolkits, the high-level view of the Graphics Scene and Actors is separated from the low-level Implementation.
Abstract classes are used to define the interfaces to the Implementations.