Add CMake commands for find_package OpenSurgSim
Add cmake commands for using OpenSurgSim from an seperate project. This
basically creates a OpenSurgSimConfig.cmake that includes the include
path and libraries of OpenSurgSim. This change currently only works for
using OpenSurgSim from the build tree, not after install.

Cmake will keep track of where OpenSurgSimConfig.cmake was located
(using the registry on windows, in ~/.cmake/packages/OpenSurgSim on
linux). So nothing else needs to be done for find_package(OpenSurgSim)
to work.