Base Mesh class, which provides basic vertex functionality using MeshVertex. Elements (edges, triangles, tetrahedra, etc.) are built with vertex IDs in a MeshElement<N>, where N is the number of vertices in the element. (See MeshElementTest.cpp for example use of edges and triangles).
MeshVertices are constructed with an optional pointer to MeshVertexData to provide extra information (see MockObjects.h and MeshTest.cpp for examples of specifying vertex data).
Similarly, MeshElements are constructed with an optional pointer to MeshElementData to provide extra information (see MockObjects.h and MestTest.cpp for examples of specifying element data).