Input Component uses a Lock Container
The input component now uses a locked container on it's data. This
renames getInputData with getData, and getData no longer returns the
data, but a data parameter must be passed in. This follows the
conventions for get on NameData and LockedContainer.

Also, added a hasData to Input Component. This is a clearer way to see
if the Input Component has gotten data yet, instead of
InputComponent->getInputData().isValid().