Add deviceConnected/deviceDisconnected to InputConsumerInterface
Added deviceConnected and deviceDisconnected callbacks to
InputConsumerInterace. These are called in CommonDevice to let the
InputConsumer know it has been connected to a device, and when it has
been disconnected. InputComponent (a subclass of
InputConsumerInterface), uses these callbacks to initialize the data
group in its locked container.

Also, now when calling InputComponent::getData, an assert makes sure
a device is connected.

This removes the need for the InputComponent::hasData. As long as a
device is connected, the input component will always have a valid
datagroup.