Add targets to EXPORT_TARGETS in subdirectories
Instead of setting CONFIG_EXPORT_TARGETS in the root CMakeLists.txt, we
setup a variable (renamed to EXPORT_TARGETS) in the root CMakeLists.txt,
and add to it whenever we install a target library. ie in the
subdirectories. This way, if a target is not compiled for some reason,
it won't mess up the OpenSurgSimConfig.