InputComponent now has a InputConsumer
Instead of the InputComponent being derived from the
InputConsumerInterface, it now has a concrete implementation of
InputConsumerInterface. This implementation (InputConsumer) puts the
input data into a locked container.

This removes the multiple inheritance of InputComponent and simplifies
the interface of InputComponent. Also, the InputComponent no longer
requires callbacks from the device to tell it when it is connected. The
InputComponent handles the connection to the device through
connectDevice/disconnectDevice and is able to record its connection
state.