Remove leading / for header install directories
Remove the need for a leading '/' for the header install directory of
surgsim_add_library. I think this makes things more clear in the usage
of surgsim_add_library. Setting the HEADER_DIRECTORY to
/SurgSim/Framework implies an absolute path in my mind,
SurgSim/Framework seems like a relative path, which it is.

Also, fixed bug where only one header file was actually getting
installed, need to add quotes to the set_target_properties call in
surgsim_add_library.