Add Physics FixedActor and its UnitTest
After discussions, we decided to have a RigidActorBase that will be the base class for any rigid motion actor (fixed actor, rigid body dynamics, rigid body with vtc...)
A state for this kind of RigidActorBase is associated to it, storing the pose of the actor.

So a FixedActor is simply an actor that does not have any physics, but you can still move it around by specifying which pose you want at any given time.
It stores differnt poses for the sake of:
1) reset (needs the initial pose)
2) future cdd (needs previous/current)