Make the OSG conversion code a bit more compact and efficient.
Use a column-major Eigen::Map of Eigen data (instead of using a
row-major map and explicitly transposing the Eigen data).

When reading, use an Eigen::Map of const data (instead of copying and
mapping non-const).

Make the templates a hair more generic, without going overboard.  This
isn't less efficient, because Eigen's cast<T>() is a no-op if T is
already the numeric type used.  (We could further reduce the number of
type coersions etc. quite a bit if we wanted to play exciting games
with templates, but I don't think that's worth the added complexity.)