Create a simple unit test for the Gauss-Seidel MLCP solver.
The MLCP solver is basically just copied from the old code, along with
a number of obsolete things it depends on.  It's incredibly smelly.
But now it's at least tested for a single input data set.

The test checks the results and the execution time, so far for a
single MLCP problem.