Work around missing std::to_string(int) in VS2010.
In VS2010, to_string() is only defined for long long, unsigned long
long, and long double types, and the compiler refuses to decide
whether an int should be promoted to signed or unsigned long long when
a call to to_string(int) is encountered.  We could explicitly cast,
but that's horribly ugly, so I just rolled our own.