Increase the Sixense start-up retry time to 5 seconds.
This is ridiculously long, but if you unplug and re-plug the Hydra, it
can take more than 4s for it to be detected.

Note that if the device has already been initialized since it was last
plugged in, it will take much less time for it to be detected (well,
1-2s, anyway).  But if no device is plugged in at all, you end up
having to wait the full 5s for an error.