Make Visual Studio understand cpplint output.
This makes errors show up in the "Error List" tab, and lets you
double-click on the error (in Error List, Output, etc) to go to the
source file.

Doing this by modifying cpplint would have been easy, but annoying for
future maintenance.  Instead, I added a wrapper script to reformat the
messages.  We can also add other checks to this wrapper.