OsgSphereActor and OsgPlaneActor.
These are quite similar. OsgSphereActor uses OsgUnitSphere which wraps an osg::Sphere shape, and OsgPlaneActor uses OsgPlane which wraps an osg::Geometry quad.

OsgUnitSphere and OsgPlane are used with SharedInstance so that the sphere and plane geodes can be shared and just scaled as needed rather than rebuilt many times.
OsgPlane approximates an infinite plane with a 1000 x 1000 quad centered at (0.0, 0.0, 0.0).