Also check the header file include guards.
The guard checking is not fully exhaustive.  For example, if a header
file contains nested namespaces "Aaa", "Bbb" and "Ccc", but the
include guard used is CCC_FILENAME_H, no warning will be reported.
(Note that you can skip any number of namespaces, but not all of them;
just FILENAME_H will produce a warning.)  This is needed to handle
pre-declarations etc:

  namespace Aaa
  {
  namespace Bbb
  {
  class Predeclared;
  };
  };
  namespace Ccc
  { ...