Major refactoring of the SixenseDevice API and related code.
SixenseDevice is now the only first-class object, and can (and should)
be created via make_shared.  When created, the device is *not*
operational; no data can be received until initialize() is called on
the device, which will happen when it is added to the device manager.
Any errors in device creation (e.g. "you didn't plug it in") will not
be reported until initialization happens.

The manager is now completely in charge of talking to the device, and
is hidden from the user-facing API.