Increase the Sixense/Hydra start-up delay.
When running in Debug with debugger attached, the very first test
after the device is plugged in takes more than 5000ms to initialize on
Dan's Asus laptop (but still reliably less than 5500).  Setting to
6000ms in case someone's hardware is slower.

The drawback is that if the device isn't plugged in, start-up takes 1s
longer to fail.