Fix (kind of) copying DLLs with CMake <= 2.8.9.
The conditional copying of DLLs etc. into the target directory based
on the build configuration (e.g. Debug vs non-Debug) is handy, but the
implementation relies on conditional generator expressions, which were
added in CMake 2.8.10.  For older versions of CMake, we now fall back
to copying those files unconditionally, with a warning message.

This is OK for now, because none of the file names we copy conflict.
In other other words, if you need FooBar_d.dll, it's OK to also copy
FooBar.dll which you don't need.  But if we really needed to support
conditional copying on older versions of CMake, we could do so using
an external tool that would be passed $<CONFIGURATION> as an argument.