Fix a very strange run-time crash on Linux.
The old code *should* be exactly equivalent to the new code, but for
some strange reason conservativeResize() of a row vector messes up the
vector's internal storage buffer.  I didn't investigate further; just
rewriting the code to use a separate temporary vector when resizing
fixed the problem.

Hurray for valgrind, by the way.