Enable Valgrind's leak checking separately from other checks.
We should move towards a leak-free code base, but many of the detected
checks seem to be triggered by the OS and/or GTest.  In the likely
case we can't do anything about those, they will need suppressions,
so I want to be able to disable leak checks for the time being.