Add core 3DConnexion support for Windows, using #ifdef.
This is currently done as a set of nasty #ifdef's right in the body of
the code.  A clean refactoring into separate files is coming up.

Also, this code initializes and runs fine on both Windows and Linux,
i.e. the visual test app behaves correctly.  But on Windows, the
device tear-down does not work correctly, so the unit tests hang.