assert when wrong type is passed to NamedAny::get
Instead of returning false, throw an assert when the NamedAny::get call
is called with the wrong datatype. I think this is the more appropriate
response.

Also, move the boost::any casting into the NamedAny::get by index call,
and call that one from the get by name function. This puts the
boost::any handling in one place.