Better mock data for custom data in DataGroups Tests
Created better mock data for the custom data in the DataGroup tests. The
new mock data is a 3D data set. It is more of an encapsulated type that better
demonstrates the intended use of custom data.

Also, added a few more tests to make sure that the NamedAny class is
behaving correclty when data isn't there, or data that is there, but
with the wrong type.