Pulled in a couple of minor Phantom/Sixense changes.
The reason for pulling them in is to make things merge cleanly in the
commits that follow.  The Phantom change actually introduces a problem
(it should say surgsim_add_library, not add_library), but that will be
fixed shortly.

These came from the feature/3dconnexion-windows branch, in the
following commit:

  commit 41a44faae4dfbdb57291e63adaeb791d6fe583e6
  Author: bert Dvornik <advornik@simquest.com>
  Date:   Wed Aug 7 14:35:50 2013 +0200

      Update existing directories that were not showing folder trees.