Fix a problem in device initialization.
Tearing down the haptics loop was failing with an assertion (and the
device wasn't producing output), because the startup wasn't creating
the haptic loop at all, because of a stupid typo in a condition
introduced during refactoring.

Thanks to @hscheirich for trying it, reporting the problem, and
working with me to debug it.