Renumber Phantom buttons as 1..4 (from 0..3).
We agreed we'd use 1-based numbering for numbered buttons as much as
possible around the time of the Sixense implementation.  The visual
test library was (re)written to follow this.  But the original Phantom
implementation was already written and using 0-based numbering.  Thank
you @hscheirich for noticing; fixed now.

Also updated the doc comments to note that the Omni and Desktop
devices have been renamed to "Geomagic Touch" and "Geomagic Touch X",
respectively.