Fix WDK detection code to work with both VS2012 and VS2010.
For VS2012, the copy of the WDK included in Windows Kit 8 (provided
with the compiler) will be detected and used.  For VS2010, the CMake
code will attempt to find a copy of WDK 7.1.0. (It may be possible to
make this work with VS2010 and WDK 8, but this has not been tested in
any way, so the architecture directory support probably isn't right.)

Note that the external WDK support relies on the fact that adding the
quoted path ".;$(IncludePath)" to the include path, using CMake's
include_directories command, works to add $(IncludePath) to the path.
This relies on some accidents of CMake's implementation [*], but we
don't know of a better way to do what we need with current CMake.

[*] CMake seems to ALWAYS prefix all include_directories() entries
    with the current directory (unless the entry is an absolute path).
    So adding "$(IncludePath)" actually passes <curdir>/$(IncludePath)
    to Visual Studio, which does not work.

    However, including an explicit semicolon passes it to Visual
    Studio without any quoting, and Visual Studio treats it as a path
    separator.  So adding ".;$(IncludePath)" (with quotes) in CMake
    passes <curdir>/.;$(IncludePath) into the Visual Studio include
    list.  Visual Studio splits this at the ';', so we end up with two
    entries: <curdir>/. (which is benign) and $(IncludePath) (which is
    what we needed).

fix #23
fix #24