Enforce a minimum lifetime for HDAL device resources.
If we destroy the callback and the device handle too soon after they
are created, the CRT complains that the HDAL accessed memory in some
horrible and improper way.  The ony way to prevent this seems to be to
wait before the objects can be destroyed.  As long as 500ms have
elapsed, things seem to behave (250ms was too short).