MassSpringState becomes TetMesh<mass, spring,void,void> and useful Eigen vectors are now in the MassSpring.
Note that this makes the class DeformableRepresentation useless as well as DeformableRepresentationState.
DeformableRepresentationState useless: each Deformable object will have its own data structure for the state now, so no more common class for that...
  Fem3D will have a state of TetrahedronElement<void, void, void, TetrahedronParameter>
  etc...
DeformableRepresentation was simply containing the 4 states, which now depends on the objects, so they cannot be hold in this base class anymore.
We could still keep this base DeformableRepresentation class for handling the setPose, setInitialPose, etc... that part should be common to all DeformableRepresentation !

Note that right now, MassSpringRep does not derive from DeformableRep, but directly from Representation, and handles the setPose, setInitialPose itself...
That could change easily...