Rename InputVtcBehavior to TransferInputPoseBehavior
InputVtcBehavior used to be typed on a VtcRigidRepresentation, but this
was not necessary, since it just does a setPose on the
VtcRigidRepresentation, which is available to all Representations. This
change generalizes this class to transfer input pose data to any
representation. Also, it adds a parameter to the behavior to change
which pose to use coming from the input. Generally this will just be
"pose", and this is the default.