Managers collect and run their own behaviors
While managers had a m_behaviors data member, they weren't collecting
their behaviors. This change adds the collection of behaviors to each
manager, so when processBehaviors is called, all the associated
behaviors are updated. Previosly, the behavior manager was collecting
all the behaviors, and only running those that had a maching manager
type of MANAGER_TYPE_BEHAVIOR.

Add processBehaviors to the doUpdate of PhysicsManager.

Move the getType from the OsgManager up to the Graphics/Manager, and
call processBehaviors in it's doUpdate instead of OsgManager.