VTC as a physics behavior
This change makes the VTC a physics behavior that controls a regular
RigidRepresentation. Previosly there was a seperate
VtcRigidRepresentatoin class that contained a lot of duplicate code with
the RigidRepresentation. This change removes alot of the duplication
by haveing the rigid representation handle the dynamics, while the VTC
just applies an external force based on the input. This more closely
follows what a VTC is, something that drives a RigidRepresentation.