VtcBehavior sends compliance to RigidRepresentation
You now need to send a compliance parameter when you call
setExternalForce and setExternalTorque. This compliance modifies the
calculated compliance matrix of the rigid representation and is used by
the VtcBehavior. This makes sense as an external spring/damper (like the
VTC) would modify the compliance of a rigid object.