Reworked RigidRepresentation::addExternalForce/Torque
First, renamed setExternalForce and setExternalTorque to
addExternalForce and addExternalTorque, respectively.

Instead of passing in a compliance matrix (which is ODE dependant and
requires unneeded inversion) we pass in a stiffness matrix (K) and
damping matrix (D) when adding a force and torque.

Becuase we are "adding" the force and torque instead of "seting" it,
this means multiple behaviors can add force/torques to the rigid
representation. However, this also means we need to zero these
force/torques after each update. So to acheive a constant force you need
to add it every physics update.