Fixed bug in intersectionsSegmentBox
Fixed bug in intersectionSegmentBox where we were incorrectly checking
segments that are parallel to the box. We were seeing if it was
parallel, beyond the minimum corner AND beyond the maximum corner. It
should be beyond the minimum cornder OR beyond the maximum corner. Added
a couple of more unit tests to check these cases. Although, this
parallel check just short circuits the calculation. If it doesn't catch
a parallel segment outside the box, the later code will catch it.

Also, renamed planeIntesections to planeIntersectionsAbsissa to be more
clear and add some comments.