Push Young modulus & Poisson ratio in base class (common to 1D/2D/3D)
All our models are defined using the linear elasticity theory (Hooke's law), and therefore all rely on these 2 parameters.
It would be good in the future to extract that to allow the use of other elasticity theory, but this is way off track to be done now (and will be a big undertaking).