Reorganized Shape and Rigid Collision Representations
ShapeCollisionRepresentation is now just a shape and a pose. It does not
contain a physics representation. This is exactly what the
MockCollisionRepresentation was, and so that is removed and
ShapeCollisionRepresentation is used intead in the Collision unit tests.

RigidCollisionRepresentation now just contains a physics representation,
and all the collision information is retreived from the physics
representation (pose, shape). Any collision representation that is
interacting with physics will need to use this RigidCollisionRepresnantaion.

The down side is that there is currently no way to specify a collision
representation with a different shape than the physics representation.
This, I think, is a seperate case, that we don't use right now, and can
be added when needed.

Conflicts:
	Examples/BouncingBalls/ExampleBouncingBalls.cpp
	SurgSim/Collision/UnitTests/BoxDoubleSidedPlaneContactCalculationTests.cpp
	SurgSim/Collision/UnitTests/BoxPlaneContactCalculationTests.cpp
	SurgSim/Collision/UnitTests/BoxSphereContactCalculationTests.cpp