Added tests to Box Capsule Contact Calculation
Added tests to make sure that contacts return normals that pointing in
the right directions (ie moving the centers apart, not closer), and that
the depths are resonable (greater than 0, and less than a maximum based
on the geometry).

These tests, actually, uncovered a bug in the code. The contact normal
wasn't pointing the correct direction, and needed to be flipped.