Renamed addNode to addData in OctreeNode
addNode was really adding data to a node at a specific position and
level. And it incidentitly would add nodes if needed. So renameing this
to addData to be more clear.

Also, added a private doAddData that will be called recursively to move
down the levels. Instead of calling the public addData. As a result, the
addData doesn't need the confusing currentLevel parameter that was only
used internally.

Added a comment to the subdivide to make it clear that subdivide in no
way 'subdivides' the data stored in the node.

Fixed some whitespace formatting.