Remove while-loop waiting for non-failed clock time in Timer::now.
The boost::steady_clock::now() call can fail, but I have not seen it fail
in the Timer class.  The while loop looked like a hack with a
magic number.  Without the while loop we might get a bad time, but that
can be checked by getting the number of clock fails.