OctreeNodes now store childrend in a std::array
Instead of stoging an octreenodes children in a std::vector, we now
store them in an array of size 8. This makes it more clear that there
are only 8 children of each node.

Also, changed isLeafNode to hasChildren. This removes any mention of
leaf node, which is really just another term for having children.

Updated the documentation and fixed some comments.

Also, fixed the testing function "countOctreeLevels" such that it makes
sure to find the maximum depth of the octree.