Remove physics elements from shape clases
To make shapes less about physics and more generally usable (as just a
geomtry), this change removes mentions of mass and inertia. Instead
shapes now work in geometric terms, ie volume, center, and the second
moment of inertia. The second central moment is just the
geometric part of the moment of inertia. ie, the moment of inertia is
the product of this second moment, and the density.

Conflicts:
	SurgSim/Math/OctreeShape-inl.h
	SurgSim/Math/UnitTests/ShapeTests.cpp