Fix temporary reference warning in Collision Representations
In Collision Representations I was returning a temporary variable as a
reference in getInitialPose. This really isn't a problem, because I call
SURGSIM_FAILURE before this line, so it never reaches the return.

However, this change reorganizes things a bit to avoid this warning.
set/getInitialPose are now implemented in RigidCollisionRepresentation
and ShapeCollisionRepresentation. And getInitialPose returns decent
values. RigidCollisionRepresentation returns the initial pose of the
underlying physics representation. ShapeCollisionRepresentation returns
its pose.