FemElement3DCube: rename some single letter variables + more doc + fix a bug
neta -> eta (that was a typo)
m_coefX -> m_shapeFucntionsXSign as these are really containing a sign (+1 or -1) and are related to shape functions
m_M -> m_mass
m_K -> m_stiffness
evaluateB ->evaluateStrainDisplacement (note that I kept the variable as B for simplicity in the equation)
evaluateJ kept as is but more details are given in the doc now. This one is hard to name differently...

Change doc style for the Shape function N because of the Latex newline \\ causing warnings under Linux.

Also fix a bug in the stress/strain evaluation. They were not integrated properly using the Gauss-quadrature rule.