Avoids the WDK include directories hack in MultiAxis' cmake, when possible.
On cmake 2.8.11+, if WDK was found, MultiAxis/CMakeLists.txt causes a
warning about Policy CMP0021.  This warning is due to a hack that adds the
Visual Studio $(IncludePath) property before the WDK include directory.
However, if the try_compile in FindWDK succeeds then WDK does not need any
include directory, and so we can avoid the hack altogether.