Added copyOctreeNode functions
Added two functions for copying octree nodes. Once for when the Data in
the OctreeNodes is the same type. In this case the data is copied over
as well. A second function is added for when the data is a different
type, in this case th data is not copied.