Makes DataGroup valid if any NamedData isValid().
DataGroup is a collection of NamedData, with constant names & indices
for the entries.  When a DataGroup is assigned via operator= (and at
other times), we assert on DataGroup::isValid(), which now returns true
if any of the NamedData isValid().

The UnitTests have been expanded to illustrate expected and incorrect
usage of DataGroup.