Renames NamedData::isValid to isInitialized.
"Valid" could have different interpretations.  The function tests
whether or not the NamedData object has an IndexDirectory, but not
the existence of anything in that IndexDirectory or the correctness of
any data associated with names/indices.  "isInitialized" is less
likely to be misinterpreted in these ways.

Also, updates documentation and assert messages for clarity.