Renames DataGroup::isValid to isInitialized.
With the renaming of NamedData::isValid to isInitialized,
DataGroup::isValid should be renamed for consistency.
Also, this avoids the possible confusion users might have about the
meaning of "valid", e.g., whether or not a DataGroup is valid if it
contains bad data.

Also updates documentation and assert messages for clarity.