Prevents OutputComponent trying to get uninitialized data from container
LockedContainer<DataGroup>::get will call DataGroup::operator=, which
asserts on the DataGroup being initialized.  So we must wait until
the DataGroup has been initialized before we try to send the data to
the device.  Since an initialized DataGroup object can only be assigned
to by a DataGroup object from which it was initialized (either copy
constructed or the first assignment after default construction), we want
to initialize the DataGroup inside the LockedContainer by the first
setData.  So don't get the data until setData has been called at least
once.