Removes CommonDevice::setInput and its single argument constructor.
CommonDevice should be passed an initial DataGroup on construction, to
setup the DataGroup that will be passed to/from the device.  By allowing
CommonDevice to be constructed without the initial DataGroup, extra
functionality was required to check whether or not that data had been
set.  The only class that used the single-argument CommonDevice
constructor was PoseIntegrator, which did so just for convenience.

Instead, PoseIntegrator now gets the DataGroup in its constructor,
which is called from MultiAxisDevice.  MultiAxisDevice gets the data
from its member RawMultiAxisDevice (which is a CommonDevice).  To do so,
MultiAxisDevice's constructor is now a friend function of
RawMultiAxisDevice.  Also, CommonDevice::getInitialData has a non-const
version.

As a bonus, this reduces the number of places in the code that we
check the validity of a DataGroup, a function which is being refactored.