Removes DataGroup::isInitialized/isValid because it is not needed.
DataGroup::isInitialized (recently renamed from isValid), was being used
in DataGroup::operator= to check if all of the member NamedData were
valid before assigning from them.  That test was not necessary because
NamedData::operator= performs the same validity check.

After removing direct calls to DataGroup::isInitialized, all of the
DataGroup's UnitTests worked without further changes, demonstrating how
the assertions are still being performed (just in NamedData).

The other places isInitialized was used were:
1) in InputComponent to test the input data (unnecessary because the
LockedContainer::set will call DataGroup::operator=), and
2) in CommonDevice's addInputConsumer and pushInput to check the
initial data and input data, respectively.

When a DataGroup is being passed between a CommonDevice and an
InputConsumerInterface, the same DataGroup object should be used for
initialization and each time pushInput is called, which should ensure
that the DataGroup's NamedData are always valid.  Also, if the
InputConsumerInterface to which the input is being pushed assigns the
DataGroup to a member variable in handleInput (like InputConsumer
setting the DataGroup into a LockedContainer), the assignment will cause
an assertion anyway (if any of the DataGroup's NamedData are not valid).