Put copying of OctreeNodes into constructors
Remove OctreeShape::fromOctreeNode, and instead create a OctreeShape
constructor that copies from a node. This constructor can handle nodes
with different templated Data. It just strips the data.

Remove copyOctreeNode functions from OctreeNode. These are replaced by
templated copy constructors in OctreeNode. If the templated Data type is
the same, the data is copied. If the Data is a different type, no data
will be copied.