Create our own gmock main to disable logging
In order to turn off logging during the unit tests, this creates our own
custom gmock main (Testing/TestingMain.cpp). This is similar to the one
included in the gmock, except it sets the default logging output
to NullOutput. Thereby disabling the logging messages that clutter up
the unit test output.

Because of this change, all the unit tests need to link against gmock as
well as gtest.