Removes unused functions and adds documentation for CommonDevice.
CommonDevice had two functions that were not being called, a non-const
version of getInitialInputData, and a const version of getInputData.
They have been removed for simplicity and to simplify refactoring of
getInputData into getter/setter.

Also, adds documentation to CommonDevice's un-documented functions, and
updates some of the documentation that was there.