Access to CommonDevice's DataGroups is now through getters/setters.
Previously, getInputData would return a reference for modification of
the DataGroup, which did not match the style throughout OSS.

The getters/setters are protected, and meant to be used by the friend
scaffolds, as described in the documentation for the functions.  This
change does not affect the PushInput/PullOutput functionality.  This is
how the scaffold will set m_inputData from the device's SDK, so that
DataGroup can be pushed to the input consumers.

Note that derived classes can access the DataGroup member variables
directly...(so can the friend scaffolds, but that looks odd).